refactor(modal): extract DB and auth helpers
